requirements.txt
Django
SQLAlchemy>=0.9.3
Flask-SQLAlchemy
dse
djangorestframework
